+++ sdk/appconfiguration/azure-data-appconfiguration/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features Added
 
 ### Breaking Changes
-
+- Change `listConfigurationSettingsForSnapshot` method's `fields` parameter type to `List`.
 ### Bugs Fixed
 
 ### Other Changes